Fix initialdata bug
